refactor(layout): split helpers.py into domain/{queries,locks,services}
Step 2b of the domain/ reorganization plan. helpers.py mixed three
distinct concerns:

* Lock factory and instance queries (need provider classes)
* Config queries that needed get_managed_lock indirectly
* Service-handler implementations

Splits into three modules along the actual dependency boundaries
the audit's queries.py grouping would have created a cycle:
coordinator.py -> domain/queries.py -> providers -> coordinator.py.

* domain/queries.py: config-only queries (no provider dependency).
  get_entry_config (preserved from #1196), get_managed_slots,
  find_entry_for_lock_slot, iter_loaded_lcm_entries (made public
  since locks.py now depends on it), get_slot_config,
  get_loaded_config_entry.

* domain/locks.py: lock-instance layer (needs BaseLock /
  INTEGRATIONS_CLASS_MAP). async_create_lock_instance,
  get_managed_locks, get_managed_lock, get_locks_from_targets.
  This file is the only one in domain/ that imports providers.

* domain/services.py: service-handler implementations.
  async_set_usercode, async_clear_usercode,
  async_set_slot_condition, async_clear_slot_condition.

Importers updated:
* __init__.py and websocket.py each split their from-helpers
  imports into queries/locks/services lines.
* tests/test_helpers.py renamed to tests/test_queries.py to track
  the canonical home of get_locks_from_targets's siblings.
* test mocks that patched helpers.INTEGRATIONS_CLASS_MAP retargeted
  to domain.locks.INTEGRATIONS_CLASS_MAP (the new import site).
