--- custom_components/lock_code_manager/helpers.py
@@ -1,242 +1,0 @@
-"""Helpers for lock_code_manager."""
-
-from __future__ import annotations
-
-from collections.abc import Iterator, Mapping
-import logging
-from typing import Any
-
-from homeassistant.components.lock import DOMAIN as LOCK_DOMAIN
-from homeassistant.config_entries import ConfigEntry, ConfigEntryState
-from homeassistant.const import (
-    ATTR_AREA_ID,
-    ATTR_DEVICE_ID,
-    ATTR_ENTITY_ID,
-    CONF_ENTITY_ID,
-)
-from homeassistant.core import HomeAssistant, callback, split_entity_id
-from homeassistant.exceptions import ServiceValidationError
-from homeassistant.helpers import (
-    config_validation as cv,
-    device_registry as dr,
-    entity_registry as er,
-)
-
-from .const import DOMAIN, EXCLUDED_CONDITION_PLATFORMS
-from .domain.queries import get_entry_config
-from .providers import INTEGRATIONS_CLASS_MAP, BaseLock
-
-_LOGGER = logging.getLogger(__name__)
-
-
-@callback
-def async_create_lock_instance(
-    hass: HomeAssistant,
-    dev_reg: dr.DeviceRegistry,
-    ent_reg: er.EntityRegistry,
-    config_entry: ConfigEntry,
-    lock_entity_id: str,
-) -> BaseLock:
-    """Generate lock from config entry."""
-    lock_entry = ent_reg.async_get(lock_entity_id)
-    assert lock_entry
-    lock_config_entry = hass.config_entries.async_get_entry(lock_entry.config_entry_id)
-    lock = INTEGRATIONS_CLASS_MAP[lock_entry.platform](
-        hass, dev_reg, ent_reg, lock_config_entry, lock_entry
-    )
-    _LOGGER.debug(
-        "%s (%s): Created lock instance %s",
-        config_entry.entry_id,
-        config_entry.title,
-        lock,
-    )
-    return lock
-
-
-def _iter_loaded_lcm_entries(hass: HomeAssistant) -> Iterator[ConfigEntry]:
-    """
-    Yield loaded Lock Code Manager config entries.
-
-    A lock may be shared by multiple LCM entries (same physical lock
-    managed from multiple Lock Code Manager configurations); callers
-    should treat the iteration as authoritative for "which locks does
-    Lock Code Manager manage right now".
-    """
-    return (
-        entry
-        for entry in hass.config_entries.async_entries(
-            DOMAIN, include_disabled=False, include_ignore=False
-        )
-        if entry.state is ConfigEntryState.LOADED
-    )
-
-
-def get_managed_locks(hass: HomeAssistant) -> dict[str, BaseLock]:
-    """
-    Return the union of locks across all loaded Lock Code Manager entries.
-
-    When two entries share the same physical lock they hold the same
-    ``BaseLock`` instance, so a flat ``entity_id -> BaseLock`` mapping is
-    well-defined.
-    """
-    return {
-        entity_id: lock
-        for entry in _iter_loaded_lcm_entries(hass)
-        for entity_id, lock in entry.runtime_data.locks.items()
-    }
-
-
-def get_locks_from_targets(
-    hass: HomeAssistant, target_data: dict[str, Any]
-) -> set[BaseLock]:
-    """Get lock(s) from target IDs."""
-    # Targets can be a single string or list; normalize for consistent iteration.
-    area_ids: list[str] = cv.ensure_list(target_data.get(ATTR_AREA_ID, []))
-    device_ids: list[str] = cv.ensure_list(target_data.get(ATTR_DEVICE_ID, []))
-    entity_ids: list[str] = cv.ensure_list(target_data.get(ATTR_ENTITY_ID, []))
-    managed_locks = get_managed_locks(hass)
-    lcm_lock_entity_ids = managed_locks.keys()
-    lock_entity_ids: set[str] = set()
-    ent_reg = er.async_get(hass)
-    lock_entity_ids.update(
-        ent.entity_id
-        for area_id in area_ids
-        for ent in er.async_entries_for_area(ent_reg, area_id)
-        if ent.domain == LOCK_DOMAIN
-    )
-    lock_entity_ids.update(
-        ent.entity_id
-        for device_id in device_ids
-        for ent in er.async_entries_for_device(ent_reg, device_id)
-        if ent.domain == LOCK_DOMAIN
-    )
-    # Split invalid (non-lock domain) from unmanaged lock entities for clearer logs.
-    invalid_entities: set[str] = set()
-    unmanaged_entities: set[str] = set()
-    for entity_id in entity_ids:
-        domain = split_entity_id(entity_id)[0]
-        if domain != LOCK_DOMAIN:
-            invalid_entities.add(entity_id)
-            continue
-        if entity_id in lcm_lock_entity_ids:
-            lock_entity_ids.add(entity_id)
-        else:
-            unmanaged_entities.add(entity_id)
-
-    if invalid_entities:
-        _LOGGER.warning(
-            "%s lock(s) are invalid lock entities: %s",
-            len(invalid_entities),
-            ", ".join(invalid_entities),
-        )
-    if unmanaged_entities:
-        _LOGGER.warning(
-            "%s lock(s) are not managed by Lock Code Manager: %s",
-            len(unmanaged_entities),
-            ", ".join(unmanaged_entities),
-        )
-
-    return {
-        lock
-        for ent_id in (lock_entity_ids & lcm_lock_entity_ids)
-        if (lock := managed_locks.get(ent_id))
-    }
-
-
-def get_managed_lock(hass: HomeAssistant, lock_entity_id: str) -> BaseLock:
-    """Get a managed lock by entity ID, raising if not found."""
-    lock = next(
-        (
-            entry.runtime_data.locks[lock_entity_id]
-            for entry in _iter_loaded_lcm_entries(hass)
-            if lock_entity_id in entry.runtime_data.locks
-        ),
-        None,
-    )
-    if not lock:
-        raise ServiceValidationError(
-            f"Lock {lock_entity_id} is not managed by Lock Code Manager"
-        )
-    return lock
-
-
-async def async_set_usercode(
-    hass: HomeAssistant, lock_entity_id: str, code_slot: int, usercode: str
-) -> None:
-    """Set a usercode on a lock slot."""
-    usercode = usercode.strip()
-    if not usercode:
-        raise ServiceValidationError(
-            "Usercode must not be empty; use the clear operation instead"
-        )
-    lock = get_managed_lock(hass, lock_entity_id)
-    await lock.async_internal_set_usercode(code_slot, usercode)
-
-
-async def async_clear_usercode(
-    hass: HomeAssistant, lock_entity_id: str, code_slot: int
-) -> None:
-    """Clear a usercode from a lock slot."""
-    lock = get_managed_lock(hass, lock_entity_id)
-    await lock.async_internal_clear_usercode(code_slot)
-
-
-def get_slot_config(config_entry: ConfigEntry, slot_num: int) -> Mapping[str, Any]:
-    """Get slot config, raising if not found."""
-    config = get_entry_config(config_entry)
-    if not config.has_slot(slot_num):
-        raise ServiceValidationError(f"Slot {slot_num} not found in config entry")
-    return config.slot(slot_num)
-
-
-def get_loaded_config_entry(hass: HomeAssistant, config_entry_id: str) -> ConfigEntry:
-    """Get a loaded config entry by ID, raising if not found or not loaded."""
-    config_entry = hass.config_entries.async_get_entry(config_entry_id)
-    if not config_entry or config_entry.domain != DOMAIN:
-        raise ServiceValidationError(
-            f"No lock code manager config entry with ID `{config_entry_id}` found"
-        )
-    if config_entry.state is not ConfigEntryState.LOADED:
-        raise ServiceValidationError(f"Config entry {config_entry.entry_id} not loaded")
-    return config_entry
-
-
-async def async_set_slot_condition(
-    hass: HomeAssistant, config_entry_id: str, slot: int, entity_id: str
-) -> None:
-    """Set a condition entity for a slot."""
-    config_entry = get_loaded_config_entry(hass, config_entry_id)
-    get_slot_config(config_entry, slot)
-
-    # Verify entity exists
-    if not hass.states.get(entity_id):
-        raise ServiceValidationError(f"Entity {entity_id} not found")
-
-    # Check for excluded platforms
-    ent_reg = er.async_get(hass)
-    entity_entry = ent_reg.async_get(entity_id)
-    if entity_entry and entity_entry.platform in EXCLUDED_CONDITION_PLATFORMS:
-        raise ServiceValidationError(
-            f"Entities from the '{entity_entry.platform}' integration are not "
-            "supported as condition entities. See the wiki for details: "
-            "https://github.com/raman325/lock_code_manager/wiki/"
-            "Unsupported-Condition-Entity-Integrations"
-        )
-
-    new_config = get_entry_config(config_entry).with_slot_field_set(
-        slot, CONF_ENTITY_ID, entity_id
-    )
-    hass.config_entries.async_update_entry(config_entry, options=new_config.to_dict())
-
-
-async def async_clear_slot_condition(
-    hass: HomeAssistant, config_entry_id: str, slot: int
-) -> None:
-    """Clear the condition entity from a slot."""
-    config_entry = get_loaded_config_entry(hass, config_entry_id)
-    get_slot_config(config_entry, slot)
-
-    new_config = get_entry_config(config_entry).with_slot_field_removed(
-        slot, CONF_ENTITY_ID
-    )
-    hass.config_entries.async_update_entry(config_entry, options=new_config.to_dict())
